requirements.txt
dspy-ai
pydantic
python-dotenv
streamlit 
